implemented new functions

+++ SciDB_analysis.py
@@ -3,6 +3,56 @@ from SciDBParallel import *
 import os, timeit, csv
 from collections import OrderedDict
 
+def ZonalStatistics(dataset, theRun):
+    """
+    This is the functions for zonal statistics
+    Each function should be completely self contained
+    """
+    
+    start = timeit.default_timer()
+    raster = ZonalStats(dataset["shape_path"], dataset["raster_path"], dataset["array_table"])
+    raster.RasterMetadata(dataset["raster_path"], dataset["shape_path"], raster.SciDBInstances, '/storage' ) 
+    stopPrep = timeit.default_timer()
+    print(raster.geoTiffPath, raster.SciDBArrayName)
+    
+    if theRun == 1:
+        datapackage = ParallelRasterization(raster.arrayMetaData, raster)
+        stopRasterization = timeit.default_timer()
+        SciDBInstances = raster.SciDBInstances
+    
+    sdb_statements = Statements(sdb)
+    
+    theAttribute = 'id:%s' % (datapackage[0]) #
+    sdb_statements.CreateLoadArray('boundary', theAttribute , 2)
+    sdb_statements.LoadOneDimensionalArray(-1, 'boundary', theAttribute, 1, 'p_zones.scidb')
+    #SciDB Load operator -1 loads in parallel
+    numDimensions = raster.CreateMask(datapackage[0], 'mask')
+    redimension_time = raster.InsertRedimension( 'boundary', 'mask', raster.tlY, raster.tlX )
+
+    summaryStatTime = raster.GlobalJoin_SummaryStats(raster.SciDBArrayName, 'boundary', 'mask', raster.tlY, raster.tlX, raster.lrY, raster.lrX, numDimensions, 1, rasterStatsCSV)
+    stopSummaryStats = timeit.default_timer()            
+    
+    timed = OrderedDict( [("connectionInfo", "XSEDE"), ("run", r), \
+        ("array_table", d["array_table"]), ("boundary_table", d["shape_path"]), ("full_time", stopSummaryStats-start), \
+        ("join_time", summaryStatTime), ("redimension_time", redimension_time), ("rasterize_time", stopRasterization-stopPrep) ])
+    sdb.query("remove(mask)")
+    sdb.query("remove(boundary)")
+    del raster
+
+    return timed
+
+def CountPixels(sdbConn, arrayTable, pixelValue):
+    """
+    This function will return the sum of the pixels in a SciDB Array
+    """
+    
+    start = timeit.default_timer()
+    query = "SELECT count(value) from %s WHERE value = %s" % (arrayTable, pixelValue)
+    sdb.aql_query(query)
+    stop = timeit.default_timer()
+    timed = OrderedDict( [("connectionInfo", "XSEDE"), ("run", r), ("analytic", "count"), ("time", stop-start) ])
+
+    return timed
 
 def datasetsprep():
     """
@@ -40,52 +90,64 @@ def WriteFile(filePath, theDictionary):
         for k in theDictionary.keys():
             theWriter.writerow(theDictionary[k])
 
+def argument_parser():
+    """
+    Parse arguments and return Arguments
+    """
+    import argparse
+
+    parser = argparse.ArgumentParser(description= "Analysis Script for running SciDB Analytics")    
+    # parser.add_argument("-Instances", required=False, nargs='*', type=int, help="Number of SciDB Instances for parallel data loading", dest="instances", default=[0,1])    
+    #parser.add_argument("-Host", required=True, help="SciDB host for connection", dest="host", default="localhost")
+    #If host = NoSHIM, then use the cmd iquery   
+    analytic = parser.add_mutually_exclusive_group(required = True)
+    analytic.add_argument('-zonal', action='store_true', dest='zonal')
+    analytic.add_argument('-count', action='store_true', dest='count')
+    analytic.add_argument('-reclassify', action='store_true', dest='reclassify')
+
+    #parser.add_argument("-r", required=True, help="Input file path for the raster", dest="rasterPath")    
+    #parser.add_argument("-a", required=True, help="Name of the destination array", dest="arrayName")
+    #parser.add_argument("-n", required=True, nargs='*', help="Name of the attribute(s) for the destination array", dest="attributes", default="value")
+    #parser.add_argument("-t", required=False, type=int, help="Size in rows of the read window, default: 8", dest="tiles", default=8)
+    #parser.add_argument("-c", required=False, type=int, help="Chunk size for the destination array, default: 1,000", dest="chunk", default=1000)
+    #parser.add_argument("-o", required=False, type=int, help="Chunk overlap size. Adding overlap increases data loading time. default: 0", dest="overlap", default=0)
+    #parser.add_argument("-TempOut", required=False, default='/home/scidb/scidb_data/0/0', dest='OutPath')
+    #parser.add_argument("-SciDBLoadPath", required=False, default='/home/scidb/scidb_data/0/0', dest='SciDBLoadPath')
+    parser.add_argument("-csv", required =False, help="Output timing results into CSV file", dest="csv", default="None")
+    parser.add_argument("-p", required =False, help="Parallel Redimensioning", dest="parallel", default="None")
+
+    return parser
 
 if __name__ == '__main__':
-
+    args = argument_parser().parse_args()
+    
     sdb = iquery()
     runs = [1,2,3]
     analytic = 1
     filePath = '/mnt/zonal_stats_3_20_2018_all.csv'
-    timings = OrderedDict()
     rasterStatsCSV = ''
 
     datasets = datasetsprep()
+    timings = OrderedDict()
+    
     for d in datasets:
-        print(d["raster_path"], d["shape_path"], d["array_table"])
         for r in runs:
-            start = timeit.default_timer()
-            
-            raster = ZonalStats(d["shape_path"], d["raster_path"], d["array_table"])
-            raster.RasterMetadata(d["raster_path"], d["shape_path"], raster.SciDBInstances, '/storage' ) 
-            stopPrep = timeit.default_timer()
-            print(raster.geoTiffPath, raster.SciDBArrayName)
-            
-            if r == 1:
-                datapackage = ParallelRasterization(raster.arrayMetaData, raster)
-                stopRasterization = timeit.default_timer()
-                SciDBInstances = raster.SciDBInstances
-            
-            sdb_statements = Statements(sdb)
-            
-            theAttribute = 'id:%s' % (datapackage[0]) #
-            sdb_statements.CreateLoadArray('boundary', theAttribute , 2)
-            sdb_statements.LoadOneDimensionalArray(-1, 'boundary', theAttribute, 1, 'p_zones.scidb')
-            #SciDB Load operator -1 loads in parallel
-            numDimensions = raster.CreateMask(datapackage[0], 'mask')
-            redimension_time = raster.InsertRedimension( 'boundary', 'mask', raster.tlY, raster.tlX )
+            if args.zonal:                  
+                print(d["raster_path"], d["shape_path"], d["array_table"])
+                timed = ZonalStatistics(d, r)
+                timings[r] = timed
+            elif args.count:
+                timed = CountPixels(sdb, d["array_table"], thePixel)
+                timings[r] = timed
 
-            summaryStatTime = raster.GlobalJoin_SummaryStats(raster.SciDBArrayName, 'boundary', 'mask', raster.tlY, raster.tlX, raster.lrY, raster.lrX, numDimensions, 1, rasterStatsCSV)
-            stopSummaryStats = timeit.default_timer()            
-            
-            timings[analytic] = OrderedDict( [("connectionInfo", "XSEDE"), ("run", r), ("SciDB_Executors", raster.SciDBInstances), ("array_table", d["array_table"]), ("boundary_table", d["shape_path"]), ("full_time", stopSummaryStats-start), ("join_time", summaryStatTime), ("redimension_time", redimension_time), ("rasterize_time", stopRasterization-stopPrep) ])
-            sdb.query("remove(mask)")
-            sdb.query("remove(boundary)")
-            del raster
-            analytic += 1
-
-        for i in range(SciDBInstances):
-            os.remove("/storage/%s/p_zones.scidb" % (i))
+            elif args.reclassify:
+                pass
+        
+        #Remove the parallel zone files after each dataset run
+        if args.zonal:      
+            for i in range(SciDBInstances):
+                os.remove("/storage/%s/p_zones.scidb" % (i))
+        
 
     if filePath: WriteFile(filePath, timings)
     print("Finished")
